Working project

+++ src/main/java/model/constants/Discount.java
@@ -2,6 +2,6 @@ package model.constants;
 
 public class Discount {
 
-    public static final double DISCOUNT = 60;
+    public static final int DISCOUNT = 60;
 
 }
